sidebar

+++ resources/views/roles/index.blade.php
@@ -1,47 +1,52 @@
 @extends('layouts.app-master')
 
 @section('content')
+    <div class="container-fluid p-5 mt-5">
+        <div class="card">
+            <div class="card-header d-flex justify-content-between text-white" style="background-color: #94489b">
+                <div>
+                    <h4>Roles</h4>
+                </div>
+                <div>
+                    <a href="{{ route('roles.create') }}" class="btn btn-warning btn-sm">Add role</a>
+                </div>
+            </div>
+            <div class="card-body table-responsive">
+                <div class="mt-2">
+                    @include('layouts.partials.messages')
+                </div>
+                <table class="table table-striped myTable">
+                    <thead>
+                        <tr>
+                            <th width="80%">Name</th>
+                            <th >Action</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @if (count($roles) > 0)
+                            @foreach ($roles as $key => $role)
+                            <tr>
+                                <td class="text-capitalize">{{ $role->name }}</td>
+                                <td>
+                                    <a class="btn btn-info btn-sm" href="{{ route('roles.show', $role->id) }}">Show</a>
+
+                                    <a class="btn btn-primary btn-sm" href="{{ route('roles.edit', $role->id) }}">Edit</a>
 
 
-    <div class="bg-light p-5 mt-5 rounded">
-        <h1>Roles</h1>
-        <div class="lead">
-            Manage your roles here.
-            <a href="{{ route('roles.create') }}" class="btn btn-primary btn-sm float-right">Add role</a>
+                                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                                    {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
+                                    {!! Form::close() !!}
+                                </td>
+                            </tr>
+                            @endforeach
+                        @endif
+
+                    </tbody>
+                </table>
+                <div class="d-flex">
+                    {!! $roles->links() !!}
+                </div>
+            </div>
         </div>
-
-        <div class="mt-2">
-            @include('layouts.partials.messages')
-        </div>
-
-        <table class="table table-bordered">
-          <tr>
-             <th width="1%">No</th>
-             <th>Name</th>
-             <th width="3%" colspan="3">Action</th>
-          </tr>
-            @foreach ($roles as $key => $role)
-            <tr>
-                <td>{{ $role->id }}</td>
-                <td>{{ $role->name }}</td>
-                <td>
-                    <a class="btn btn-info btn-sm" href="{{ route('roles.show', $role->id) }}">Show</a>
-                </td>
-                <td>
-                    <a class="btn btn-primary btn-sm" href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                </td>
-                <td>
-                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                    {!! Form::close() !!}
-                </td>
-            </tr>
-            @endforeach
-        </table>
-
-        <div class="d-flex">
-            {!! $roles->links() !!}
-        </div>
-
     </div>
 @endsection
